+++ api/app/hilo_simulator.go
@@ -4,6 +4,7 @@ import (
 	"github.com/Z-M-Huang/Tools/api"
 	"github.com/Z-M-Huang/Tools/data"
 	"github.com/Z-M-Huang/Tools/data/apidata/application"
+	"github.com/Z-M-Huang/Tools/data/constval"
 	"github.com/Z-M-Huang/Tools/utils"
 	hilosimulator "github.com/Z-M-Huang/hilosimulator"
 	"github.com/gin-gonic/gin"
@@ -11,7 +12,7 @@ import (
 
 //HILOSimulate /api/hilo-simulator/simulate
 func HILOSimulate(c *gin.Context) {
-	response := c.Keys[utils.ResponseCtxKey].(*data.Response)
+	response := c.Keys[constval.ResponseCtxKey].(*data.Response)
 	request := &application.HiLoSimulateRequest{}
 
 	err := c.ShouldBind(&request)
@@ -82,7 +83,7 @@ func HILOSimulate(c *gin.Context) {
 
 //HILOVerify /api/hilo-simulator/verify
 func HILOVerify(c *gin.Context) {
-	response := c.Keys[utils.ResponseCtxKey].(*data.Response)
+	response := c.Keys[constval.ResponseCtxKey].(*data.Response)
 	request := &application.HiLoVerifyRequest{}
 
 	err := c.ShouldBind(&request)
