+++ data/constVal/constVal.go
@@ -1,0 +1,14 @@
+package constval
+
+const (
+	//ClaimCtxKey claim context key
+	ClaimCtxKey string = "claim_context_key"
+	//ResponseCtxKey page data context key
+	ResponseCtxKey string = "response_context_key"
+	//SessionCookieKey auth token in session key
+	SessionCookieKey string = "session_token"
+	//UsedTokenKey contains apps used. This is a slice of app titles
+	UsedTokenKey string = "used_token"
+	//PageStyleCookieKey allows users to choose the style wants to use for page
+	PageStyleCookieKey string = "page_style"
+)
